# drivers: spi_dw_flash: read performance optimizations