--- a/boards/tenstorrent/tt_blackhole/tt_blackhole_tt_blackhole_smc.dts
+++ b/boards/tenstorrent/tt_blackhole/tt_blackhole_tt_blackhole_smc.dts
@@ -29,7 +29,7 @@
 		compatible = "snps,designware-spi-flash";
 		reg = <0>;
 		status = "disabled";
-		spi-max-frequency = <DT_FREQ_M(3)>;
+		spi-max-frequency = <DT_FREQ_M(9)>;
 	};
 };