--- a/samples/drivers/flash/performance/boards/tt_blackhole_tt_blackhole_smc.overlay
+++ b/samples/drivers/flash/performance/boards/tt_blackhole_tt_blackhole_smc.overlay
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2025 Tenstorrent AI ULC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <mem.h>
+
+&spi_flash {
+	status = "okay";
+	partitions {
+		compatible = "fixed-partitions";
+		#address-cells = <1>;
+		#size-cells = <1>;
+
+		storage_partition: partition@2000000 {
+			label = "storage";
+			reg = <0x2000000 DT_SIZE_K(64)>;
+		};
+	};
+};