--- a/samples/drivers/flash/performance/prj.conf
+++ b/samples/drivers/flash/performance/prj.conf
@@ -1,0 +1,1 @@
+CONFIG_FLASH=y