… samples/drivers/flash/performance/boards/tt_blackhole_tt_blackhole_smc.conf