When reading in 4 byte aligned increments, we can optimize the performance of the SPI DW driver by using a data frame size of 32 bytes (this optimization is also possible for 16 byte frames if desired).

Implement this performance optimization within the SPI DW flash driver. In addition, raise the clock frequency of the SPI controller, as we can now read faster without FIFO overruns.

Add a flash performance sample, to help validate the performance of the SPI DW driver (and test for FIFO overflows)

Output of flash performance sample without this change (SPI clocked at 2MHz to pass read without FIFO overflow)
```
*** Booting Zephyr OS build v4.1.0-rc3 ***
Erasing 8 pages at 0x2000000
Erase took 116 ms
Write of 32768 bytes took 257 ms
Read of 32768 bytes took 15 ms
Flash performance test complete
```

Output of flash performance sample with this change (SPI clocked at 9MHz):
```
*** Booting Zephyr OS build v4.1.0-rc3 ***
Erasing 8 pages at 0x2000000
Erase took 116 ms
Write of 32768 bytes took 222 ms
Read of 32768 bytes took 3 ms
Flash performance test complete
```