@danieldegrasse 2025-03-11
This optimization is primarily useful on the read side, since we will need fast read support for MCUBoot to function well