@danieldegrasse 2025-03-12
> I'm good with https://github.com/tenstorrent/tt-zephyr-platforms/pull/10, as long as you retest https://github.com/tenstorrent/tt-zephyr-platforms/pull/3. We used to have issues making octal mode DDR/DQS works with frame size 32, that why we only use frame size 8. But you might be able to fix it.

Great- I'll merge this first, as it won't affect the core app. Then can retest #3 and validate octal/quad mode in the bootrom there.